pyproject.toml
[project]
name = "multimodality_app"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.21",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.3",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.0",
    "rich>=14.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
